Fix `associate` get_completion_signatures
