+++ src/beman/execution/unstoppable_scheduler.cppm
@@ -1,0 +1,11 @@
+module;
+// src/beman/execution/unstoppable_scheduler.cppm                       -*-C++-*-
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
+#include <beman/execution/detail/unstoppable_scheduler.hpp>
+
+export module beman.execution.detail.unstoppable_scheduler;
+
+namespace beman::execution::detail {
+export using beman::execution::detail::unstoppable_scheduler;
+} // namespace beman::execution::detail
